Issue #DBC-2023 feat: Protected Content Read
